extract <ChannelList>
